Add init script for brother daemon from blog. Update logging and merge other sanity patches for Brother MFC scanning.
git-svn-id: http://svn.freemedsoftware.org/freemed-utilities/scan_import/trunk@99 0cc4c890-d802-0410-860e-a516f263db2e

+++ convert_and_import_pdf.sh
@@ -21,10 +21,23 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 
-TS=$(date +%s)
 ORIG="$1"
+TS=$(date +%s)
 
+if [ ! -r "${ORIG}" ]; then
+	echo "Could not open file '${ORIG}'";
+	logger -t convert_and_import_pdf "Could not open file '${ORIG}', exiting."
+	exit 1
+fi
+
+logger -t convert_and_import_pdf "Converting ${ORIG} into djvu image"
 djvudigital --dpi=200 "${ORIG}" "${TS}.$(basename "${ORIG}").djvu" 2>&1
+if [ -f "${ORIG}.txt" ]; then
+	logger -t convert_and_import_pdf "Importing text into ${TS}.$(basename "${ORIG}").djvu"
+	djvused "${TS}.$(basename "${ORIG}").djvu" -e "set-txt ${ORIG}.txt"
+fi
+logger -t convert_and_import_pdf "Importing ${TS}.$(basename "${ORIG}").djvu"
 "$(dirname "$0")/import_djvu.pl" "${TS}.$(basename "${ORIG}").djvu"
 rm -f "${ORIG}"
+logger -t convert_and_import "Completed convert/import for ${ORIG}"
